Improved path handling for imports
Slightly improved path handling for the import builtin, non-absolute string handling will be fixed in the future.
